Allow and add skeleton for exercise project
The user submission `submission.jl` will be copied to this project folder
`/exercise_project/` during grading. The file needs to be named
`submission.jl` for this to work, since it is included by the project root
file `src/exercise.jl`.

The teacher needs to provide the file `runtests.jl`, which should be copied to
`/exercise_project/test/runtests.jl` by `run.sh`. This will allow the project
tests to be run. This might be refined further, so that all tests do not need
to reside in `runtests.jl`. The container itself could also take care of the
copying.

The script `run.sh` also needs to `cd("/exercise_project") before running any
Julia tests via Pkg.test().